requirements.bot_principal.txt
httpx
pybit
python-dotenv